requirements.txt
streamlit
opencv-python
numpy
scikit-image
matplotlib
Pillow
